requirements.txt
asgiref==3.6.0
aspose-words==23.1.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
confection==0.0.4
cymem==2.0.7
Django==4.1.6
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.5.0/en_core_web_md-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
idna==3.4
Jinja2==3.1.2
langcodes==3.3.0
lxml==4.9.2
MarkupSafe==2.1.2
murmurhash==1.0.9
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
Pillow==9.4.0
preshed==3.0.8
pydantic==1.10.4
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.7.1
requests==2.28.2
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.5
thinc==8.1.7
tqdm==4.64.1
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.14
wasabi==1.1.1
